+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-28T06:21:39.943Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-28T09:00:06.548Z")
     r"""The date and time the resource was last updated"""