+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWARVNWDK578FH1AY1ABVNMC"
+        "01JWB1XSRV9BVVTE5JTX12FVNC"
     )
     r"""The id of the resource"""