+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWARVNWGNX72T58MD4DAYAZG"
+        "01JWB1XSRWWVWM796XCPWA5T35"
     )
     r"""The id of the resource"""